refactor rate api clients to a separate package
