+++ pkg/rateclient/coingeckorate_test.go
@@ -1,7 +1,7 @@
-package core_test
+package rateclient_test
 
 import (
-	"bitcoinrateapp/pkg/core"
+	"bitcoinrateapp/pkg/rateclient"
 	"context"
 	"fmt"
 	"io"
@@ -28,7 +28,7 @@ func TestValueRequest(t *testing.T) {
 	expectedRate := 1000.0
 
 	client := &MockHTTPClient{expectedRate: expectedRate}
-	coingecko := core.NewCoingeckoRateWithHTTPClient("bitcoin", "uah", client)
+	coingecko := rateclient.NewCoingeckoRateWithHTTPClient("bitcoin", "uah", client)
 	actualRate, err := coingecko.Value(context.TODO(), "bitcoin", "uah")
 	if err != nil {
 		t.Error(err)
