+++ pkg/rateclient/rate.go
@@ -1,0 +1,7 @@
+package rateclient
+
+type Rate interface {
+	Value() float64
+	Coin() string
+	Currency() string
+}
