+++ pkg/testenv/mockrate.go
@@ -1,8 +1,8 @@
 package testenv
 
 import (
-	"bitcoinrateapp/pkg/core"
 	"bitcoinrateapp/pkg/model"
+	"bitcoinrateapp/pkg/rateclient"
 	"context"
 )
 
@@ -10,6 +10,6 @@ type MockRate struct {
 	ExpectedRate float64
 }
 
-func (m *MockRate) Value(_ context.Context, _, _ string) (core.Rate, error) {
+func (m *MockRate) Value(_ context.Context, _, _ string) (rateclient.Rate, error) {
 	return model.NewExchangeRate(m.ExpectedRate, "coin", "currency"), nil
 }
